+++ source/lib/tests/test_env_mat_a.cc
@@ -504,11 +504,12 @@ TEST_F(TestEnvMatA, prod_cpu) {
   deepmd::InputNlist inlist(nloc, &ilist[0], &numneigh[0], &firstneigh[0]);
   deepmd::convert_nlist(inlist, nlist_a_cpy);
 
-  std::vector<double> em(nloc * ndescrpt), em_deriv(nloc * ndescrpt * 3),
-      rij(nloc * nnei * 3);
-  std::vector<int> nlist(nloc * nnei);
-  std::vector<double> avg(ntypes * ndescrpt, 0);
-  std::vector<double> std(ntypes * ndescrpt, 1);
+  std::vector<double> em(static_cast<size_t>(nloc) * ndescrpt),
+      em_deriv(static_cast<size_t>(nloc) * ndescrpt * 3),
+      rij(static_cast<size_t>(nloc) * nnei * 3);
+  std::vector<int> nlist(static_cast<size_t>(nloc) * nnei);
+  std::vector<double> avg(static_cast<size_t>(ntypes) * ndescrpt, 0);
+  std::vector<double> std(static_cast<size_t>(ntypes) * ndescrpt, 1);
   deepmd::prod_env_mat_a_cpu(&em[0], &em_deriv[0], &rij[0], &nlist[0],
                              &posi_cpy[0], &atype_cpy[0], inlist, max_nbor_size,
                              &avg[0], &std[0], nloc, nall, rc, rc_smth, sec_a);
@@ -538,11 +539,12 @@ TEST_F(TestEnvMatA, prod_cpu_equal_cpu) {
   std::vector<int *> firstneigh(nloc);
   deepmd::InputNlist inlist(nloc, &ilist[0], &numneigh[0], &firstneigh[0]);
   convert_nlist(inlist, nlist_a_cpy);
-  std::vector<double> em(nloc * ndescrpt), em_deriv(nloc * ndescrpt * 3),
-      rij(nloc * nnei * 3);
-  std::vector<int> nlist(nloc * nnei);
-  std::vector<double> avg(ntypes * ndescrpt, 0);
-  std::vector<double> std(ntypes * ndescrpt, 1);
+  std::vector<double> em(static_cast<size_t>(nloc) * ndescrpt),
+      em_deriv(static_cast<size_t>(nloc) * ndescrpt * 3),
+      rij(static_cast<size_t>(nloc) * nnei * 3);
+  std::vector<int> nlist(static_cast<size_t>(nloc) * nnei);
+  std::vector<double> avg(static_cast<size_t>(ntypes) * ndescrpt, 0);
+  std::vector<double> std(static_cast<size_t>(ntypes) * ndescrpt, 1);
   deepmd::prod_env_mat_a_cpu(&em[0], &em_deriv[0], &rij[0], &nlist[0],
                              &posi_cpy[0], &atype_cpy[0], inlist, max_nbor_size,
                              &avg[0], &std[0], nloc, nall, rc, rc_smth, sec_a);
