+++ source/op/soft_min_virial.cc
@@ -73,7 +73,8 @@ class SoftMinVirialOp : public OpKernel {
     OP_REQUIRES(context, (nloc == du_tensor.shape().dim_size(1)),
                 errors::InvalidArgument("number of du should match"));
     OP_REQUIRES(context,
-                (nloc * nnei * 3 == sw_deriv_tensor.shape().dim_size(1)),
+                (static_cast<int64_t>(nloc) * nnei * 3 ==
+                 sw_deriv_tensor.shape().dim_size(1)),
                 errors::InvalidArgument("number of sw_deriv should match"));
     OP_REQUIRES(context,
                 (static_cast<int64_t>(nloc) * nnei * 3 ==
